Create Wiki table and associations
A new table that we'll use to refactor assumptions that all entities live on
enwiki.

All existing and new data is linked with the default {language}wiki--this is a
transitional arrangement, until the UI allows us to specify a target wiki.

Native_id is where we'll store the onwiki IDs for article, revision, and user.
These fields will default to the legacy .id until we can break assumptions.

We create aliases like :rev_id to make it easier to import native IDs from the
wikis.

TODO:
* Course and User wiki are not specified well enough to implement yet.