Reference users by username rather than by id
The transitional native_id strategy is not used here, because the situation with
users is complicated by having a different local id on each wiki.  Rather than
deal with that, we go directly to a global username.  global_id might be better
still, but would require changes to the Replica backend.

Make the user_exists? test more obvious.